Small type changes

+++ chemicals/virial.py
@@ -223,7 +223,7 @@ def Z_from_virial_density_form(T, P, *args):
     solns = np.roots(arr)
     for rho in solns:
         if abs(rho.imag) < 1e-12 and rho.real > 0.0:
-            return P/(R*T*rho.real)
+            return float(P/(R*T*rho.real))
     raise ValueError("Could not find real root")
